requirements.txt
Spark
Kafka
Elasticsearch